+++ qlit/simple.py
@@ -148,11 +148,6 @@ class SimpleThesaurus():
         termset = self.t.get_related(ref)
         return SimpleTerm.from_termset(termset)
 
-    def get_all(self) -> list[SimpleTerm]:
-        """All terms as dicts."""
-        terms = self.simple_terms.values()
-        return sorted(terms, key=lambda term: term['prefLabel'].lower())
-
     def search(self, s: str) -> Termset:
         """Find terms matching a user-given incremental (startswith) search string."""
         scored_hits = self.searcher.search(s)